fix: move AWSServiceConfiguration Platform extension
